Move UnsupportedFormatError from encoders to errors module
This is a breaking change for how to import the UnsupportedFormatError class,
but it makes more sense logically to have all the error classes defined in the
errors module.
