+++ src/sagemaker_containers/_errors.py
@@ -12,6 +12,8 @@
 # language governing permissions and limitations under the License.
 from __future__ import absolute_import
 
+import textwrap
+
 import six
 
 
@@ -54,3 +56,14 @@ class ExecuteUserScriptError(_CalledProcessError):
 class ChannelDoesNotExistException(Exception):
     def __init__(self, channel_name):
         super(ChannelDoesNotExistException, self).__init__('Channel %s is not a valid channel' % channel_name)
+
+
+class UnsupportedFormatError(Exception):
+    def __init__(self, content_type, **kwargs):
+        self.message = textwrap.dedent(
+            """Content type %s is not supported by this framework.
+
+            Please implement input_fn to to deserialize the request data or an output_fn to
+            serialize the response. For more information, see the SageMaker Python SDK README."""
+            % content_type)
+        super(Exception, self).__init__(self.message, **kwargs)
